[fix] 🐿️ WhilickItem이 바뀌어도 sound on/off 설정한 것 유지되도록 수정사항 WhilickPage에 globalAudioState 생성
